+++ tests/test_repository.py
@@ -144,6 +144,36 @@ class TestBitBucketRepository:
         ]
         assert mock.await_args[0][0][: len(expected_cmd)] == expected_cmd
 
+    async def test_token_with_username_added_correctly_from_credential(
+        self, monkeypatch
+    ):
+        """Ensure that the repo url is in the format `https://username:token@url`."""  # noqa: E501
+
+        class p:
+            returncode = 0
+
+        mock = AsyncMock(return_value=p())
+        monkeypatch.setattr(prefect_bitbucket.repository, "run_process", mock)
+        credential = "XYZ"
+        username = "ABC"
+        repo = "https://bitbucket.org/PrefectHQ/prefect.git"
+        b = BitBucketRepository(
+            repository=repo,
+            bitbucket_credentials=BitBucketCredentials(
+                token=SecretStr(credential), username=username
+            ),
+        )
+        await b.get_directory()
+        assert mock.await_count == 1
+        expected_cmd = [
+            "git",
+            "clone",
+            f"https://{username}:{credential}@bitbucket.org/PrefectHQ/prefect.git",
+            "--depth",
+            "1",
+        ]
+        assert mock.await_args[0][0][: len(expected_cmd)] == expected_cmd
+
     async def test_ssh_fails_with_credential(self, monkeypatch):
         """Ensure that credentials cannot be passed in if the URL is not in the HTTPS
         format.