+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `get_client` method and `url` field in `BitbucketCredentials` - [#7](https://github.com/PrefectHQ/prefect-bitbucket/pull/7/)
+- support to log into Bitbucket Server with `token` and `username` - [#11](https://github.com/PrefectHQ/prefect-bitbucket/pull/11/)
 
 ### Changed